More redis tests, at diff redis types
